Removed classes that uses urls and BlockLifeSpans from project because they were moved to the extractor project.
